Fix locking
I ran into a deterministic test failure, "No route matches", when two
particular tests ran; the first was a regular integration test[^1], and
the second was an AC[^2] integration test.

I assumed that this was an issue with parallel tests (which Rails has
provided by default since 6.0), which led me to believe that I needed a
RW lock[^3] on all tests. I wanted to allow regular tests to run in
parallel while forcing AC tests to be isolated and run in serial.

I implemented it, and while doing so, I moved the position of the lock
so that the entire test was locked, instead of only the filesystem
writes. But I continued seeing the same error. I tried various things
and found that adding another `Rails.application.reload_routes!` fixed
the issue.

The latter fix led me to wonder if the RW lock had any effect at all.
So I switched to a regular lock that locked the entire AC test and found
that it continued to work.

That then led me to wonder if the lock needed to be changed at all. It
didn't.

[^1]: A "regular" is any test that doesn't involve anonymous controllers
[^2]: An AC test is an anonymous controller test
[^3]: A readers-writer lock, where there could be multiple simultaneous
readers but only one writer. When the writer is active, there can be no
other readers.